Conflicts:
	web/.gitignore
	web/app/Config/bootstrap.php
	web/app/Config/routes.php
	web/app/Controller/AppController.php
	web/app/Controller/EventsController.php
	web/app/Model/AppModel.php
	web/app/Model/Event.php
	web/app/View/Events/index.ctp
	web/app/View/Events/view.ctp
	web/app/View/Layouts/ajax.ctp
	web/app/View/Layouts/default.ctp
	web/app/webroot/css/cake.generic.css

+++ src/zm_buffer.cpp
@@ -43,9 +43,9 @@ unsigned int Buffer::expand( unsigned int count )
     int headSpace = mHead - mStorage;
     int tailSpace = spare - headSpace;
     int width = mTail - mHead;
-    if ( spare > count )
+    if ( spare > (int)count )
     {
-        if ( tailSpace < count )
+        if ( tailSpace < (int)count )
         {
             memmove( mStorage, mHead, mSize );
             mHead = mStorage;
